refactor agendaItemSubscription into teamSubscription
