+++ src/server/graphql/types/RemoveAgendaItemPayload.ts
