+++ src/universal/mutations/UpdateAgendaItemMutation.ts
@@ -1,4 +1,5 @@
 import {commitMutation, graphql} from 'react-relay'
+import {RecordProxy} from 'relay-runtime'
 import Atmosphere from 'universal/Atmosphere'
 import handleUpdateAgendaItems from 'universal/mutations/handlers/handleUpdateAgendaItems'
 import {IUpdateAgendaItemOnMutationArguments} from 'universal/types/graphql'
@@ -7,7 +8,7 @@ import updateProxyRecord from 'universal/utils/relay/updateProxyRecord'
 import {UpdateAgendaItemMutation} from '__generated__/UpdateAgendaItemMutation.graphql'
 
 graphql`
-  fragment UpdateAgendaItemMutation_agendaItem on UpdateAgendaItemPayload {
+  fragment UpdateAgendaItemMutation_team on UpdateAgendaItemPayload {
     agendaItem {
       id
       isComplete
@@ -22,10 +23,15 @@ const mutation = graphql`
       error {
         message
       }
-      ...UpdateAgendaItemMutation_agendaItem @relay(mask: false)
+      ...UpdateAgendaItemMutation_team @relay(mask: false)
     }
   }
 `
+
+const updateAgendaItemUpdater = (payload: RecordProxy, {store}) => {
+  const teamId = payload.getValue('teamId')
+  handleUpdateAgendaItems(store, teamId)
+}
 
 const UpdateAgendaItemMutation = (
   atmosphere: Atmosphere,
@@ -38,7 +44,9 @@ const UpdateAgendaItemMutation = (
     mutation,
     variables,
     updater: (store) => {
-      handleUpdateAgendaItems(store, teamId)
+      const payload = store.getRootField('updateAgendaItem')
+      if (!payload) return
+      updateAgendaItemUpdater(payload, {store})
     },
     optimisticUpdater: (store) => {
       const proxyAgendaItem = store.get(updatedAgendaItem.id)
